Add new ApiServiceRegistrations for registering API services
