+++ appBackend/common-service-data/src/jsMain/kotlin/com/gchristov/thecodinglove/commonservicedata/api/ApiRequest.kt
@@ -17,23 +17,3 @@ inline fun <reified T> ApiRequest.bodyAsJson(
 interface ApiParameterMap {
     operator fun <T> get(key: String): T?
 }
-
-class ApiRequestFacade(private val parametersMapFacade: ApiParametersMapFacade) {
-    operator fun invoke(request: FirebaseFunctionsHttpsRequest): ApiRequest = object : ApiRequest {
-        override val headers: ApiParameterMap = parametersMapFacade(request.headers)
-
-        override val query: ApiParameterMap = parametersMapFacade(request.query)
-
-        override val body: Any = request.body as Any
-
-        override val rawBody: String = request.rawBody
-    }
-}
-
-class ApiParametersMapFacade {
-    operator fun invoke(
-        parameterMap: FirebaseFunctionsHttpsParameterMap
-    ): ApiParameterMap = object : ApiParameterMap {
-        override fun <T> get(key: String): T? = parameterMap[key]
-    }
-}
